Trigger IDocumentIndexHandler events from NamedIndexingService
